refactor(services): share cache time/encoding helpers
now_secs, fresh, and the hand-rolled hex encoder were byte-identical copies in tidal/cache.rs and sportify/cache.rs. Moves them to services/cache_util.rs so the two caches can't drift on clock or TTL semantics.

hash_query stays per-module: the inputs genuinely differ (Sportify keys also fold in the search kind), so there is nothing to share there.

Adds unit tests for the shared helpers (TTL boundary, clock-skew via future fetched_at, hex encoding). 631 tests pass; both caches' existing round-trip and expiry tests still green.
